+++ packages/boot/src/components/save-panel/index.tsx
@@ -1,15 +1,15 @@
 /**
  * WordPress dependencies
  */
-import { useState } from '@wordpress/element';
-import { Modal } from '@wordpress/components';
-import { EntitiesSavedStates } from '@wordpress/editor';
+import { useState, Suspense } from '@wordpress/element';
+import { Modal, Spinner } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
  */
 import useSaveShortcut from './use-save-shortcut';
+import LazyEntitiesSavedStates from './lazy-entities-saved-states';
 
 export default function SavePanel() {
 	const [ isOpen, setIsOpen ] = useState< boolean >( false );
@@ -26,10 +26,12 @@ export default function SavePanel() {
 			title={ __( 'Review changes' ) }
 			size="small"
 		>
-			<EntitiesSavedStates
-				close={ () => setIsOpen( false ) }
-				variant="inline"
-			/>
+			<Suspense fallback={ <Spinner /> }>
+				<LazyEntitiesSavedStates
+					close={ () => setIsOpen( false ) }
+					variant="inline"
+				/>
+			</Suspense>
 		</Modal>
 	);
 }